Multi-step install support for 10.x images and s1 image support for pr1f

+++ nx-os/poap/poap.py
@@ -1277,6 +1277,9 @@ def is_image_cs_or_msll():
     if (os.path.exists("/isan/etc/noncs.txt")):
         return 1
 
+    if (os.path.exists("/isan/etc/s1.txt")):
+        return 3
+
     return 0
     
 def target_system_image_is_currently_running():
@@ -1306,6 +1309,8 @@ def target_system_image_is_currently_running():
             image_parts64.insert(0, "nxos64-cs")
         elif is_cs == 1:
             image_parts64.insert(0, "nxos64-msll")
+        elif is_cs == 3:
+            image_parts64.insert(0, "nxos64-s1")
         else: 
             image_parts64.insert(0, "nxos64")
         
@@ -1508,12 +1513,15 @@ def install_nxos_issu():
        up with. 
     ''' 
     if global_copy_image:
-        system_image_path = os.path.join(options["destination_path"],
-                                     options["destination_system_image"])
+        if multi_step_install == True:
+            system_image_path = os.path.join(options["destination_path"],options["destination_midway_system_image"])
+        else: 
+            system_image_path = os.path.join(options["destination_path"],options["target_system_image"])
+
         system_image_path = system_image_path.replace("/bootflash", "bootflash:", 1)
     else:
         system_image_path = os.path.join("bootflash:",get_currently_booted_image_filename())
-    
+
     try:
         os.system("touch /tmp/poap_issu_started")
         poap_log("terminal dont-ask ; install all nxos %s no-reload non-interruptive" % system_image_path)
@@ -2312,7 +2320,7 @@ def set_next_upgrade_from_user():
     else:
         poap_log("Already on the midway image, upgrading to %s next." %
                  options["target_system_image"])
-
+        multi_step_install = False
 
 def set_next_upgrade_from_upgrade_path():
     """Checks the currently running image and the target image to see where on the upgrade path
@@ -2592,8 +2600,6 @@ def main():
     # the directory structure needed, if any
     create_destination_directories()
 
-    if (len(options["install_path"]) != 0 and options["mode"] != "personality"):
-        parse_poap_yaml()
     check_multilevel_install()
     # In two step install we just copy the midway image and reboot.
     # Config copy and script download happens in the second step.
@@ -2604,16 +2610,18 @@ def main():
         download_scripts_and_agents()
         # End of multi_step_install is False block
 
-    if (len(options["install_path"]) != 0 and options["mode"] != "personality"):
-        validate_yaml_file()
-        copy_poap_files()
-        time.sleep(2)
-        install_license()
-        install_rpm()
-        time.sleep(2)
-        install_certificate()
-        time.sleep(2)
-        copy_standby_files()
+        if (len(options["install_path"]) != 0 and options["mode"] != "personality"):
+            parse_poap_yaml()
+            poap_log("Starting the RPM copy and installation section. ")
+            validate_yaml_file()
+            copy_poap_files()
+            time.sleep(2)
+            install_license()
+            install_rpm()
+            time.sleep(2)
+            install_certificate()
+            time.sleep(2)
+            copy_standby_files()
         
     copy_system()
 
@@ -2641,10 +2649,11 @@ def main():
         cli('copy bootflash:%s scheduled-config' % options["split_config_first"])
         poap_log("Done copying the first scheduled cfg")
         remove_file("/bootflash/%s" % options["split_config_first"])
+    if multi_step_install == False:
+        cli('copy bootflash:%s scheduled-config' % options["split_config_second"])
+        poap_log("Done copying the second scheduled cfg")
+        remove_file(os.path.join("/bootflash", options["split_config_second"]))
 
-    cli('copy bootflash:%s scheduled-config' % options["split_config_second"])
-    poap_log("Done copying the second scheduled cfg")
-    remove_file(os.path.join("/bootflash", options["split_config_second"]))
     if (options["use_nxos_boot"] == False):
         install_nxos_issu()
